requirements.txt
streamlit==1.16.0
openai==0.27.0
langchain==0.0.148
python-dotenv==1.0.0
faiss-cpu==1.7.3
